fix(format): use portable format specifiers for int64_t/uint64_t
On LP64 (64-bit Linux/macOS), int64_t is long (8 bytes), not long long.
Using %lld for int64_t arguments causes -Wformat warnings because the
format specifier expects unsigned long long but the argument type is long.

Fix by casting int64_t arguments to (long long) and uint64_t to
(unsigned long long) at format string call sites. Also fix:
- PDSCLIDispatcher: cast data.bytes to (const char *) for %s
- WebSocketServer: use gai_strerror(gai_err) instead of raw int for %s

+++ Garazyk/Sources/Sync/Firehose/SubscribeReposHandler.m
@@ -905,7 +905,7 @@ static void *kSubscribeReposEventQueueKey = &kSubscribeReposEventQueueKey;
       if (patchedData) {
         eventData = patchedData;
         PDS_LOG_SYNC_DEBUG(@"Patched stale seq %lld -> %lld in replayed %@ event",
-                            [payload[@"seq"] longLongValue], dbSeq, msgType);
+                            [payload[@"seq"] longLongValue], (long long)dbSeq, msgType);
       }
     }
